+++ templates/admin/tab-shortcodes.php
@@ -3,9 +3,17 @@
         <tbody>
         <tr>
             <td>
-                <textarea name="" id="main-shortcode" readonly
+                <textarea name="" readonly
                           onclick="this.select()">[anycomment include="true"]</textarea>
                 <p class="description"><?php echo __( 'Use this shortcode do display comment box in custom place (e.g. landing page).', 'anycomment' ) ?></p>
+            </td>
+        </tr>
+
+        <tr>
+            <td>
+                <textarea name="" readonly
+                          onclick="this.select()">[anycomment_socials]</textarea>
+                <p class="description"><?php echo __( 'Use this shortcode do display list of available social networks.', 'anycomment' ) ?></p>
             </td>
         </tr>
         </tbody>